Few tweaks for themeing
This is just a handful of small tweaks that I had to make over the
course of building the bdbq elqastiquill theme. Notably it replaces the
page URLs with a separate "posts" theme, so the landing page can have a
hero story and banner and such, but not interfere with older content
pages.